+++ modules/apps/foundation/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/internal/tag/facet/constants/TagFacetPortletKeys.java
@@ -12,14 +12,14 @@
  * details.
  */
 
-package com.liferay.portal.search.web.internal.type.facet.constants;
+package com.liferay.portal.search.web.internal.tag.facet.constants;
 
 /**
- * @author André de Oliveira
+ * @author Lino Alves
  */
-public class TypeFacetWebKeys {
+public class TagFacetPortletKeys {
 
-	public static final String DISPLAY_CONTEXT =
-		"TYPE_FACET_PORTLET_DISPLAY_CONTEXT";
+	public static final String TAG_FACET =
+		"com_liferay_portal_search_web_tag_facet_portlet_TagFacetPortlet";
 
 }