+++ portal-impl/src/com/liferay/portal/servlet/filters/aggregate/AggregateFilter.java
@@ -18,10 +18,12 @@ import com.liferay.portal.kernel.configuration.Filter;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.servlet.BrowserSniffer;
+import com.liferay.portal.kernel.servlet.BrowserSnifferUtil;
 import com.liferay.portal.kernel.servlet.BufferCacheServletResponse;
 import com.liferay.portal.kernel.servlet.HttpHeaders;
 import com.liferay.portal.kernel.servlet.PortalWebResourceConstants;
 import com.liferay.portal.kernel.servlet.PortalWebResourcesUtil;
+import com.liferay.portal.kernel.servlet.ResourceUtil;
 import com.liferay.portal.kernel.servlet.ServletResponseUtil;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.CharPool;
@@ -139,6 +141,12 @@ public class AggregateFilter extends IgnoreModuleRequestFilter {
 						urlConnection.getInputStream());
 				}
 				else {
+					int queryPos = importFileName.indexOf(CharPool.QUESTION);
+
+					if (queryPos != -1) {
+						importFileName = importFileName.substring(0, queryPos);
+					}
+
 					ServletPaths importFileServletPaths = servletPaths.down(
 						importFileName);
 
@@ -335,9 +343,7 @@ public class AggregateFilter extends IgnoreModuleRequestFilter {
 			return null;
 		}
 
-		String requestURI = request.getRequestURI();
-
-		String resourcePath = requestURI;
+		String resourcePath = request.getRequestURI();
 
 		String contextPath = request.getContextPath();
 
@@ -355,14 +361,11 @@ public class AggregateFilter extends IgnoreModuleRequestFilter {
 					resourcePath.substring(pos);
 		}
 
-		URL resourceURL = _servletContext.getResource(resourcePath);
+		URL resourceURL = ResourceUtil.getResourceURL(
+			resourcePath, request.getRequestURI(), _servletContext);
 
 		if (resourceURL == null) {
-			resourceURL = PortalWebResourcesUtil.getResource(resourcePath);
-
-			if (resourceURL == null) {
-				return null;
-			}
+			return null;
 		}
 
 		String cacheCommonFileName = getCacheFileName(request);
@@ -374,7 +377,8 @@ public class AggregateFilter extends IgnoreModuleRequestFilter {
 
 		if (cacheDataFile.exists() &&
 			(cacheDataFile.lastModified() >=
-				URLUtil.getLastModifiedTime(resourceURL))) {
+				URLUtil.getLastModifiedTime(resourceURL)) &&
+			!_isLegacyIe(request)) {
 
 			if (cacheContentTypeFile.exists()) {
 				String contentType = FileUtil.read(cacheContentTypeFile);
@@ -397,7 +401,9 @@ public class AggregateFilter extends IgnoreModuleRequestFilter {
 
 			response.setContentType(ContentTypes.TEXT_CSS);
 
-			FileUtil.write(cacheContentTypeFile, ContentTypes.TEXT_CSS);
+			if (!_isLegacyIe(request)) {
+				FileUtil.write(cacheContentTypeFile, ContentTypes.TEXT_CSS);
+			}
 		}
 		else if (resourcePath.endsWith(_JAVASCRIPT_EXTENSION)) {
 			if (_log.isInfoEnabled()) {
@@ -447,22 +453,9 @@ public class AggregateFilter extends IgnoreModuleRequestFilter {
 
 	protected String getCssContent(
 		HttpServletRequest request, HttpServletResponse response,
-		String resourcePath, String content) {
+		ServletContext cssServletContext, String resourcePath, String content) {
 
 		try {
-			ServletContext cssServletContext = null;
-
-			String requestURI = request.getRequestURI();
-
-			if (PortalWebResourcesUtil.hasContextPath(requestURI)) {
-				cssServletContext =
-					PortalWebResourcesUtil.getPathServletContext(requestURI);
-			}
-
-			if (cssServletContext == null) {
-				cssServletContext = _servletContext;
-			}
-
 			content = DynamicCSSUtil.replaceToken(
 				cssServletContext, request, content);
 		}
@@ -490,23 +483,57 @@ public class AggregateFilter extends IgnoreModuleRequestFilter {
 	}
 
 	protected String getCssContent(
+		HttpServletRequest request, HttpServletResponse response,
+		String resourcePath, String content) {
+
+		try {
+			String requestURI = request.getRequestURI();
+
+			ServletContext cssServletContext =
+				ResourceUtil.getPathServletContext(
+					resourcePath, requestURI, _servletContext);
+
+			return getCssContent(
+				request, response, cssServletContext, resourcePath, content);
+		}
+		catch (Exception e) {
+			_log.error("Unable to detect servlet context " + resourcePath, e);
+
+			if (_log.isDebugEnabled()) {
+				_log.debug(content);
+			}
+
+			response.setHeader(
+				HttpHeaders.CACHE_CONTROL,
+				HttpHeaders.CACHE_CONTROL_NO_CACHE_VALUE);
+
+			return content;
+		}
+	}
+
+	protected String getCssContent(
 			HttpServletRequest request, HttpServletResponse response,
 			URL resourceURL, String resourcePath)
 		throws IOException {
 
-		ServletContext cssServletContext = null;
 		String resourcePathRoot = null;
 
 		String requestURI = request.getRequestURI();
 
-		if (PortalWebResourcesUtil.hasContextPath(requestURI)) {
-			cssServletContext = PortalWebResourcesUtil.getPathServletContext(
-				requestURI);
-			resourcePathRoot = "/";
-		}
+		ServletContext cssServletContext = ResourceUtil.getPathServletContext(
+			resourcePath, requestURI, _servletContext);
 
-		if (cssServletContext == null) {
-			cssServletContext = _servletContext;
+		if (PortalWebResourcesUtil.hasContextPath(requestURI)) {
+			resourcePathRoot = PortalWebResourcesUtil.stripContextPath(
+				cssServletContext, resourcePath);
+
+			resourcePathRoot = ServletPaths.getParentPath(resourcePathRoot);
+
+			if (resourcePathRoot.equals(StringPool.BLANK)) {
+				resourcePathRoot = "/";
+			}
+		}
+		else {
 			resourcePathRoot = ServletPaths.getParentPath(resourcePath);
 		}
 
@@ -514,10 +541,16 @@ public class AggregateFilter extends IgnoreModuleRequestFilter {
 
 		String content = StringUtil.read(urlConnection.getInputStream());
 
+		if (_isLegacyIe(request)) {
+			return getCssContent(
+				request, response, cssServletContext, resourcePath, content);
+		}
+
 		content = aggregateCss(
 			new ServletPaths(cssServletContext, resourcePathRoot), content);
 
-		return getCssContent(request, response, resourcePath, content);
+		return getCssContent(
+			request, response, cssServletContext, resourcePath, content);
 	}
 
 	protected String getJavaScriptContent(URL resourceURL) throws IOException {
@@ -564,6 +597,16 @@ public class AggregateFilter extends IgnoreModuleRequestFilter {
 				ServletResponseUtil.write(response, (String)minifiedContent);
 			}
 		}
+	}
+
+	private boolean _isLegacyIe(HttpServletRequest request) {
+		if (BrowserSnifferUtil.isIe(request) &&
+			(BrowserSnifferUtil.getMajorVersion(request) < 10)) {
+
+			return true;
+		}
+
+		return false;
 	}
 
 	private static final String _BASE_URL = "@base_url@";